+++ client/src/data-provider/Agents/index.ts
@@ -1,1 +1,2 @@
 export * from './queries';
+export * from './mutations';
